Fixed the deadlock possibility. Probably. Hopefully. Found some other problems and fixed them as well.
There was actually a second deadlock condition discovered while examining and modifying the code for this. Both were "masked" by the "band-aid" of just increasing the two ArrayBlockingQueue message inbox's buffer size.

=====

The first identified deadlock condition was when the main PeerProcess process received messages from ClientThreads, where the individual ClientThread had obtained a file piece (forwarded from its ServerThread partner). After writing in the file piece data and updating the bitfield, etc, the ClientThread sends a message to the PeerProcess, asking the PeerProcess to "broadcast" commands to all the ClientThreads, telling each ClientThread to send a "have piece" message tot its TCP partner. If a single ClientThread's message inbox was full, however, it could not receive this command, and the PeerProcess would block, waiting for the ClientThread's inbox to free up space. However, the ClientThread could technically have received another file data piece, and it would need to tell the PeerProcess to broadcast another round of "have piece" for that new piece. If the PeerProcess's inbox was full, then the ClientThread would block. Thus, both the PeerProcess and the ClientThread would block, waiting for each other to empty their inbox.

Essentialy this first deadlock was an issue of circular message passing between the ClientThread and PeerProcess. This was solved by, instead of having the ClientThread send the "please tell all ClientThreads to broadcast have-piece messages" to the PeerProcess, to instead have the ServerThread that originally received the file data piece to send that message to the PeerProcess. Thus, the flow of messages is: 1) ServerThread -> ClientThread, 2) ServerThread -> PeerProcess, 3) PeerProcess -> ClientThread. This was, the circular message passing was eliminated, and any "full inbox" blocking would be eventually cleared by the ClientThread processing it's inbox.

=====

The second identified deadlock was found during subsequent testing made after that change. This one was caused by the use of synchronization locks.

When the main PeerProcess does not have any messages to process, it examines the timers for the preferred and optimistic choking. In this routine, it obtains the "peerProcessLock" synchronization lock, which is used to prevent the PeerProcess and the multiple ClientThreads from modifying state information, such as bytes downloaded and whether a peer had the complete file or not.

During checking of new unchoking peer partners, the PeerProcess would send messages to its ClientThreads, telling them to send "choke" or "unchoke" messages through their handled TCP connections. PeerProcess tells the ClientThreads to do this task, via adding messages to the ClientThreads' individual inboxes. If the particular ClientThread's inbox was full, however, the PeerProcess would block - while still holding the "peerProcessLock" hostage.

Some of the ClientThread's inbox tasks require that "peerProcessLock", so the ClientThread would block and never empty its inbox.

This was solved by, instead of having the PeerProcess try to add the tasks to the ClientThread's inboxes while holding the lock, to instead maintain a queue. Every time the PeerProcess would have tried to add an inbox message, it instead adds that task to its internal queue. At the end of the unchoking routine, when the syncrhonization lock was released, then PeerProcess would process this internal queue, and send the tasks to the ClientThreads. This way, the ClientThreads could obtain that "peerProcessLock" lock, and make space in their inboxes.

=====

During examination of code for ClientThread.java, a redundant "determine whether I am interested in the neighbor peer or not" routine method-call was found in the code for when the ClientThread is processing "send have piece" messages out through the TCP connection.

The interest-state should not change when just sending a "have piece" message. The interest-state WILL change for when processing "neighbor has piece" messages, however. For that code, the "re-determine interest" routine call was correctly present.

So I removed the "re-determine interest" call from the "send I have piece" message routine. That call is still there for the "receieve neighbor has piece" message, though.

======

During testing that the program still worked after all these above changes, a rare bug was found. During one particular test, all processes except one successfully terminated.

Upon examination of the one process that didn't terminate, it was found that it had downloaded all file pieces EXCEPT for one.

What had happened was that a peer received a piece (Piece # 1466 or something). The ServerThread correctly received the piece, added a task message to the ClientThread to tell it to update the bitfield, and then added a task message to the PeerProcess to tell it to add tasks messages to all the ClientThreads to send "have piece" messages.

The ClientThread correctly processed the bitfield and write-file-data-out command first, and then processed the "send have piece" task it got subsequently from the PeerProcess.

One of its neighbors, after having received the "have piece" message, requested that piece.

However, when the request for the piece was processed, a check of that bitfield returned "false", which would indicate that the piece was not present to respond to the request with. So that one piece was never sent.

However, the piece had actually been received previously (or else a "have piece" message that led to the subsequent request for that piece would not have been sent out in the first place). Basically, it should have been impossible for a request of a piece to be received, while not having that piece.

My guess was that this was due to the check of the bitfield not being synchronized. My guess is that reading of BitSet's bitfields is not thread-safe, even when the particular index of the BitSet can not possibly be written to concurrently (although the BitSet object as a whole may be being written concurrently). So, while the Bitset's true value was "I have the piece", it may have erronenously returned "I don't have the piece" since an other thread may have writing to the BitSet object at the same time, although it would be writing to an other piece's index, as other previously-implemented syncrhonization locks would make it impossible to write to a piece index that is currently being read.

Therefore, I added "clientThreadLock" syncrhonization do the problematic "sendPiece()" method of ClientThread.java . With this change made, all references to the "myPeer" bitfield, even if they are just reading operation, all synchronized, which I hope fixes that rare bug that appeared.

=====

After all these changes, I made the two message inboxes (in PeerProcess.java and ClientThread.java) have a queue buffer size of just "1". Testing three times with the large "tree.jpg" image and six peers it worked correctly (although obviously slower due to waiting for inboxes to have free space, at like 5 minutes long versus the normal 2-3 minutes).

=====

I also tested an edge case where the "Common.cfg" may have specification that there are 0 preferred peers, meaning that there would only be the 1 optimistic peer allowed. The program still worked and terminated correctly (obviously slower to complete though).

=====

After these tests, I reverted the two message inbox sizes to what they were previously (999 for PeerProcess.java and 9,999 for ClientThread.java).