Pushed around new problems caused by fixing the deadlock like three times (new solutions make new problems, and the solutions to those made more problems!) until it seems to work correctly now...
Note bottom part of this update about run time speed getting worse (3 minutes to 8 minutes), and the "fix" for that (1 minute!).

=====

Peers were getting requests for pieces that they had not yet processed (written file piece out and updated bitfield). This was because the previous change, that made it so that the ServerThread informs the PeerProcess to broadcast "have piece" messages, created the possibility of a neighbor peer being informed of the "myPeer" having the piece, even though it had not yet been processed by the ClientThread that is supposed to process it.

For example, after the changes made before, the program may function as follows: For a "myPeer" of 1001, ServerThread 1002 may receieve piece #999 from Peer 1002. ServerThread then adds a message to ClientThread 1002's inbox, telling it to process that piece. After adding that message, ServerThread then tells PeerProcess to command the ClientThreads to broadcast "have piece" messages. PeerProcess does just that, and adds messages to all the ClientThread's inboxes, telling them to send "have piece" messages. Although ClientThread 1002 will do the correct ordering of first processing the piece before sending a "have piece" message (since it got the message from ServerThread 1002 before PeerProcess's broadcast message), the other ClientThreads are not constrained to wait for the processing of the piece. For example, ClientThread 1003 can send that "have piece" message right away, and Peer 1003 may then request Piece 999. Peer 1003 sends a request message to ServerThread 1003, which forwards that to ClientThread 1003. At this point, if ClientThread 1002 has not yet processed the original piece #999 (for example, if ClientThread 1002 is backed up and has other work in the inbox before it reached that process piece message), then ClientThread 1003 will try to send piece #999, even though it hasn't been processed (written and bitfield updated) yet.

The fix for this was to add a lock, at the beginning step when ServerThread 1002 receives the original piece #999. ServerThread 1002 will now create a "blocking lock" using an "on-the-fly generated" BlockingArrayQueue with capacity of 1, and then immediately filling that capacity with whatever junk message (integer value of 0 for this case). This BlockingArrayQueue lock will be sent along with the ThreadMessage to both ClientThread 1002, which is supposed to process the piece, and to the PeerProcess, which is supposed to broadcast commands to all the ClientThreads to send "have piece".

When ClientThread 1002 gets around to processing the piece, it will remove the junk value entry from the queue. Meanwhile, PeerProcess will attempt to add a junk value entry to the queue. Since the queue was made to have capacity of only 1 (and was prefilled with a value by the ServerThread), PeerProcess can't progress until ClientThread 1002 processes the piece and frees up the queue. Only when the queue is freed, can PeerProcess progress and broadcast commands to the ClientThreads to send "have piece" messages.

=====

On another note, yet another problem was made by the fix for the OTHER deadlock condition. The other deadlock's fix was solved by having PeerProcess "accumulate" choke/unchoke commands to ClientThreads for peers that were choked or unchoked in the routine "every X seconds" unchoking evaluation. Before, the deadlock problem was that these choke/unchoke commands were trying to be added to the ClientThread's inbox while the PeerProcess held a lock (since it needed to inspect peers' state information on the number of bytes downloaded and whether they had the full file or not). The fix was as described, to delay this sending of messages until the lock was released.

However, the new problem arose in that, once the lock is released, the choke/unchoke messages may not be immediately added to the ClientThread's inboxes like they were previously. This new behavior caused ClientThreads to sometimes eventually make an "request piece evaluation" decision to request the same piece from the same peer. The code was not written to take this possibility into account. The original code had an if condition to check if the requested piece was a duplicate request, but did not have an else condition to actually act if a duplicate request was received (at the time it was known to be impossible to receive a duplicate request).

Certain state variables need to be cleared upon receiving a request (or else some process would wait forever for the state variables to be cleared), and this was only performed in the if condition. The fix was to have those state variables be cleared out in the else condition (when duplicate requests are received) as well.

=====

Also related to the above problem was the new problem that the pieces would actually be sent multiple times. This caused duplicate "got x piece from y" messages to be logged in the "official" peer logs.

The fix for this was to check if the piece was obtained previously, and if so, to just ignore the duplicate receiving of the piece.

The problem of the pieces being sent multiple times sometimes still persists, but does not cause deadlock or incorrectness of the download, and with this fix, the official log appears correct.

To fix the problem of the pieces being sent multiple times would mean reverting the previous changes made to fix the deadlock issues from before... so I guess I'll leave it as it is now.

=====

Also reverted the previous GitHub upload's removal of the "re-determine interest" call from the "send I have piece" message routine. That re-determine interest call actually made sense, since the "send I have piece" message routine is performed eventually after the peer receives the piece. When the routine receives the piece, it does not do the "re-determine interest" there, but instead delays it to when the "send I have piece" message is sent. So the original inclusion of that was correct. The code would still function with the deletion made before, however the behavior would technically not be correct to the project's specifications. Therefore, reverted that previous deletion.

=====

After all these changes, I performed basically the same tests again on my localhost computer, using the "tree.jpg" 23 MB image with the 6 peers.

1) "Normal" test with PeerProcess queue of capacity 999 and ClientThread queue capacity of 9,999

2) Test with both inbox capacities reduced to 1.

3) Test with both inbox capacities reduced to 1 AND "Common.cfg" modified to have 0 max preferred neighbors (so only the optimistic neighbor can be unchoked)

4) PeerProcess queue of capacity 999 and ClientThread queue capacity of 9,999  AND "Common.cfg" modified to have 0 max preferred neighbors (so only the optimistic neighbor can be unchoked)

5) "Normal" test (again) with PeerProcess queue of capacity 999 and ClientThread queue capacity of 9,999

---> They all seemed to terminate correctly and have the image files transferred correctly. Run time varied from like 8 minutes to 15 minutes (for the reduced queue and reduced preferred peers cases).

This was markedly worse than the "normal" 3 minute-ish run time from before. This was likely due to me adding like triple the amount of debugging print-outs than before (basically printing out when each synchronization lock was trying to be obtained, when it was actually obtained, and when it was released).

I then commented out all the debugging System.out.print codes (I missed a few), and the performance went straight up to 1 minute when doing a "normal" test.

=====

Hopefully the program stays working! But it seems like alot of these bugs were only discovered "some of the times", so who knows if these 6 localhost tests were enough to prove lack-of-bugs...