Changed peerProcess.java and PeerObject.java so that "FileWriter" would not be created for peers that are not the single peer that the program was called for.
Before, peer_# folders would be created for all peers, and empty data files would be created in those folders. This meant that a peer that was downloading a file had folders for all peers, even though it only needs the folder for itself.

Now, with this change, each peer will only make a data folder (e.g. peer_1001) for itself.